Add an e2e test directory for using rules_fish with bzlmod.
To support this, all build targets were moved from rules_fish into this sub-workspace.
That includes porting out the repo and toolchain from the root `MODULE.bazel` and
copying down boilerplate to define a workspace.

To enable running these e2e tests with local changes in the repo, we use
[local_path_override](https://bazel.build/rules/lib/globals/module#local_path_override)
which required marking the workspace boundaries with `WORKSPACE.bazel` files.
This also required ignoring the sub-workspace with a `.bazelignore` file specifying
that path.